Moved queries to dZoneManager
Moved the queries added in this PR to dZoneManager so they arent called so often.  Fixed variable typing of m_maxLevel as well.

+++ dGame/dComponents/CharacterComponent.h
@@ -259,7 +259,7 @@ public:
 	 */
 	void UpdatePlayerStatistic(StatisticID updateID, uint64_t updateValue = 1);
 
-    uint64_t GetMaxLevel() {return m_MaxLevel;};
+    uint32_t GetMaxLevel() { return m_MaxLevel; };
 
     /**
      * Character info regarding this character, including clothing styles, etc.
@@ -291,7 +291,7 @@ private:
     /**
      * Maximum level of the entity
      */
-    uint64_t m_MaxLevel;
+    uint32_t m_MaxLevel;
 
     /**
      * Universe score of the entity
